requirements.txt
Django==5.1.5
djangorestframework==3.15.2
pymongo==4.12.0
python_bcrypt==0.3.2
